requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pypdf==4.0.1
python-docx==1.1.0
spacy==3.7.2
jinja2==3.1.3
python-dotenv==1.0.1
pydantic-settings==2.1.0
